% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_wrappers.R
\name{cor_wrap}
\alias{cor_wrap}
\title{cor_wrap}
\usage{
cor_wrap(x, y, method = "pearson", what = c("all", "cor", "conf", "pval"), ...)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{method}{Which correlation to compute.
Takes as special value 'chatterjee' to compute revised
Chatterjee correlation coefficient (no p values returned in that case)}

\item{...}{Other parameters passed to \code{cor.test}}

\item{return}{Whether to return only the correlation coefficient, the confidence interval, the p value, or all}
}
\value{
what If all, a named list, otherwise a numeric vector
}
\description{
Wrapper for cor.test with controlled returns
}
\author{
Jesko Wagner
}
