% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIM.R
\name{XIMcalculate}
\alias{XIMcalculate}
\title{XIMcalculate}
\usage{
XIMcalculate(xvec, yvec, M)
}
\arguments{
\item{xvec}{Vector}

\item{yvec}{Vector}

\item{M}{the number of right nearest neighbors.}
}
\value{
the value of \verb{$\\xi_\{n,M\}$}.
}
\description{
This function computes the revised Chatterjee’s rank correlation \verb{$\\xi_\{n,M\}$}.
}
\author{
Zhexiao Lin and Fang Han
}
