% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIM.R
\name{XIMsim}
\alias{XIMsim}
\title{XIMsim}
\usage{
XIMsim(n, M, B)
}
\arguments{
\item{n}{number of samples}

\item{M}{the number of right nearest neighbors.}

\item{B}{the number of simulation.}
}
\value{
the value of \verb{$\\xi_\{n,M\}^\{\\pm\}$}.
}
\description{
This function computes the simulation statistics \verb{$\\xi_\{n,M\}^\{\\pm(b)\}$}.
}
\author{
Zhexiao Lin and Fang Han
}
