% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIM.R
\name{XIMstat}
\alias{XIMstat}
\title{XIMstat}
\usage{
XIMstat(xvec, yvec, M)
}
\arguments{
\item{xvec}{Vector}

\item{yvec}{Vector}

\item{M}{the number of right nearest neighbors.}
}
\value{
the value of \verb{$\\xi_\{n,M\}^\{\\pm\}$}.
}
\description{
This function computes the test statistics \verb{$\\xi_\{n,M\}^\{\\pm\}$}.
}
\author{
Zhexiao Lin and Fang Han
}
