% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIM.R
\name{XIMtest}
\alias{XIMtest}
\title{XIMtest}
\usage{
XIMtest(xvec, yvec, M, B, alpha = 0.05)
}
\arguments{
\item{xvec}{Vector}

\item{yvec}{Vector}

\item{M}{is the number of right nearest neighbors}

\item{B}{is the number of simulation}

\item{alpha}{is the signiﬁcance level.}
}
\value{
a logical value, TRUE reject, FALSE accept.
}
\description{
This function performs the simulation based test directly.
}
\author{
Zhexiao Lin and Fang Han
}
