% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIM.R
\name{XIMtestT}
\alias{XIMtestT}
\title{XIMtestT}
\usage{
XIMtestT(XIMstat, XIMsim, alpha = 0.05)
}
\arguments{
\item{XIMstat}{is the value of \verb{$\\xi_\{n,M\}^\{\\pm\}$}}

\item{XIMsim}{is the vector of \verb{$\\xi_\{n,M\}^\{\\pm(b)\}$}}

\item{alpha}{is the signiﬁcance level.}
}
\value{
a logical value, TRUE reject, FALSE accept.
}
\description{
This function performs the simulation based test using test statistics and simulation statistics.
}
\author{
Zhexiao Lin and Fang Han
}
